Fix error when installing an addon that will not be registered
Some addons are removed from the plugin registration lists. These
include addons that don't meet their prerequisite requirements
and also developer tools and unstable addons when not running in
developer mode.

In the current design we don't store the installation status of
these addons, so this fix just presents an error dialog to the
user.

A better long-term solution would be to register the addons, but
flag them as inactive.  This way we could easily detect if that
they were installed and disable the install button.

Fixes #13233.